Switched gears; this is Markdown-only.
The original idea I had was to make a literate programming tool that was not
just code-agnostic, but markup agnostic, too. This resulted in a pretty
complicated parser and confusing, complex regular expression requirements.

It basically worked in that I had a self hosting project in reStructuredText
and, simply by changing the parser config, started using a Markdown file. I
don't recall what the reStructuredText configuration looked like, but this is
what I was using while I converted it to the current Markdown-only markup.

    #{code_end => "^\n[\\S]",
      code_start => "^(?<consume>\n###### (?<tag>.+)\n)",
      macro => "###### (?<name>.+)",
      escaped_macro => "^(#)_(#####)"}

That `escaped_macro` thing is stupid. The current implementation escapes macros
by prefixing it with a `\` which is normal and allows easy nesting of escaping.
The underscore thing was a kludge.

It's possible that now that I understand the problem domain a little better
that I will come back and try and make it markup agnostic again. However, I now
think that my original goal was actually *only* about web versus book
publishing. Markdown was meant, according to its author, to produce HTML and
that's all I want.

(I used to be all about reStructuredText but have changed my mind -- there's a
reason the tool support is weak for that markup. The grammar is complex and
docutils itself is a very (complex) XML/OOP thingummy.)