Switch to Pandoc-style fenced code.
Indented code is no longer supported. I didn't want to support both styles
because I think the fenced code is almost always desirable.

With fenced code, you can add metadata to the code block for syntax
highlighting. Knot now uses [Pandoc's fenced code attributes][]. This style

    ```{.erlang name="my code section"}
    foo() -> ok.
    ```

...renders a syntax-highlighted Erlang code block in GitHub's and Pandoc's
output. (It doesn't work in Bitbucket, but their Markdown output is
aesthetically weak anyway.)

The name attribute is used internally by knot to name code sections, but it
also shows up as a name attribute on the HTML node in Pandoc's output. This
will be useful if I attempt to write front end code for a source code browser.

This commit actually has two steps. At one point the document had the
`######`-style code block names and also the new name attribute. This was so
knot could compile it's next version. I was going to commit it in two stages,
but I had a nebulous grasp of it in my mind. I just went forward with the
conversion.

The fact that I was able to read, understand, and migrate this code is proof
that this literate progamming is working. I love it. I'm even writing documents
as I explore code at work. Communicating is wonderful.

[Pandoc's fenced code attributes]: http://pandoc.org/README.html#extension-fenced_code_attributes

fixes #7